+++ src/Products/PloneMeeting/MeetingConfig.py
@@ -2590,7 +2590,7 @@ schema = Schema((
                 'update_local_roles': SelectColumn(
                     "Labels config update local roles?",
                     col_description="labels_config_update_local_roles_col_description",
-                    vocabulary="listBooleanVocabulary",
+                    vocabulary_factory="ConfigLabelsConfigUpdateLocalRoles",
                     default='0'),
             },
             label='Labelsconfig',
@@ -3839,7 +3839,7 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
     security.declareProtected(WriteRiskyConfig, 'setUsingGroups')
 
     def setUsingGroups(self, value, **kwargs):
-        '''Overrides the field 'setUsingGroups' mutator to enable or disable
+        '''Overrides the field 'usingGroups' mutator to enable or disable
            the MEETING_REMOVE_MOG_WFA WFA when relevant.
            Updating WF role mappings and every meetings local_roles is managed
            by the onConfigModified event.'''
@@ -3864,6 +3864,22 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
             # value changed, need to update local roles but WFA is already selected
             self.REQUEST.set('need_update_%s' % MEETING_REMOVE_MOG_WFA, True)
         self.getField('usingGroups').set(self, value, **kwargs)
+
+    security.declareProtected(WriteRiskyConfig, 'setFolderTitle')
+
+    def setFolderTitle(self, value, **kwargs):
+        '''Overrides the field 'folderTitle' mutator to rename every member folder
+           title when folderTitle changed.'''
+        stored = self.getField('folderTitle').get(self, **kwargs)
+        if stored != value:
+            folders = self._get_all_meeting_folders()
+            extras = 'number_of_elements={0} MeetingConfig={1} old_folder_title={2} new_folder_title={3}'.format(
+                len(folders), self.getId(), stored, value)
+            fplog('update_members_folder_title', extras=extras)
+            for folder in folders:
+                folder.setTitle(value)
+                folder.reindexObject(idxs=['Title'])
+        self.getField('folderTitle').set(self, value, **kwargs)
 
     security.declarePublic('getUsedVoteValues')
 
@@ -4635,8 +4651,8 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
         if len(label_ids) != len(set(label_ids)):
             return translate(
                 'labels_config_can_not_have_several_config_for_same_label',
-                 domain='PloneMeeting',
-                 context=self.REQUEST)
+                domain='PloneMeeting',
+                context=self.REQUEST)
 
     security.declarePrivate('validate_customAdvisers')
 
@@ -6255,6 +6271,7 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
         ))
         return res
 
+
     def listCommitteesEnabled(self):
         '''Vocabulary for committees.enabled datagrid column.'''
         d = "PloneMeeting"
@@ -6284,6 +6301,7 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
                     confidential_profiles.append('{0}{1}'.format(CONFIGGROUPPREFIX, po_infos['row_id']))
             else:
                 confidential_profiles.append('{0}{1}'.format(READERPREFIX, suffix))
+        fct_titles = {fct['fct_id']: fct['fct_title'] for fct in get_registry_functions(as_copy=False)}
         for suffix in get_item_validation_wf_suffixes(self):
             confidential_profiles.append('{0}{1}'.format(PROPOSINGGROUPPREFIX, suffix))
 
@@ -6293,14 +6311,14 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
             if profile.startswith(PROPOSINGGROUPPREFIX):
                 res.append(
                     (profile,
-                     translate('visible_for_{0}'.format(PROPOSINGGROUPPREFIX),
-                               mapping={'meeting_group_suffix':
-                                        translate(profile.replace(PROPOSINGGROUPPREFIX, ''),
-                                                  domain="PloneMeeting",
-                                                  context=self.REQUEST)},
-                               domain="PloneMeeting",
-                               context=self.REQUEST,
-                               default=u"Visible for {0}".format(profile))))
+                     translate(
+                        'visible_for_{0}'.format(PROPOSINGGROUPPREFIX),
+                        mapping={
+                            'meeting_group_suffix':
+                                fct_titles[profile.replace(PROPOSINGGROUPPREFIX, '')]},
+                        domain="PloneMeeting",
+                        context=self.REQUEST,
+                        default=u"Visible for {0}".format(profile))))
             elif profile.startswith(CONFIGGROUPPREFIX):
                 config_group_suffix = profile.replace(CONFIGGROUPPREFIX, '')
                 is_power_observer = config_group_suffix in po_row_ids
@@ -6471,7 +6489,8 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
             expression=tal_expr,
             roles_bypassing_expression=[],
             extra_expr_ctx=extra_expr_ctx,
-            empty_expr_is_true=empty_expr_is_true)
+            empty_expr_is_true=empty_expr_is_true,
+            raise_on_error=True)
 
     def getItemIconColorName(self):
         '''This will return the name of the icon used for MeetingItem portal_type.'''
@@ -6614,6 +6633,8 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
                     portalType.icon_expr_object = Expression(portalType.icon_expr)
                     catalog = api.portal.get_tool('portal_catalog')
                     brains = catalog.unrestrictedSearchResults(portal_type=portal_type)
+                    extras = 'number_of_elements={0} portal_type={1}'.format(len(brains), portalType.id)
+                    fplog('update_item_icon_color', extras=extras)
                     pghandler = ZLogHandler(steps=1000)
                     pghandler.init(
                         'Updating items icon color ({0})...'.format(metaTypeName), len(brains))
@@ -8145,14 +8166,16 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
         """Return every meeting folders for this MeetingConfig."""
         folders = []
         portal = api.portal.get()
-        for userFolder in portal.Members.objectValues():
-            mymeetings = getattr(userFolder, 'mymeetings', None)
-            if not mymeetings:
-                continue
-            meetingFolder = getattr(mymeetings, self.getId(), None)
-            if not meetingFolder:
-                continue
-            folders.append(meetingFolder)
+        members = portal.get('Members')
+        if members:
+            for userFolder in members.objectValues():
+                mymeetings = getattr(userFolder, 'mymeetings', None)
+                if not mymeetings:
+                    continue
+                meetingFolder = getattr(mymeetings, self.getId(), None)
+                if not meetingFolder:
+                    continue
+                folders.append(meetingFolder)
         return folders
 
     def _synchSearches(self, folder=None):