+++ src/Products/PloneMeeting/profiles/__init__.py
@@ -525,10 +525,10 @@ class MeetingConfigDescriptor(Descriptor):
             {'edit_access_on': '',
              'edit_access_on_cache': '1',
              'edit_groups': [
-                'configgroup_meetingmanagers',
-                'suffix_proposing_group_creators',
-                'suffix_proposing_group_prereviewers',
-                'suffix_proposing_group_reviewers'],
+                 'configgroup_meetingmanagers',
+                 'suffix_proposing_group_creators',
+                 'suffix_proposing_group_prereviewers',
+                 'suffix_proposing_group_reviewers'],
              'edit_groups_excluding': '0',
              'edit_states': [],
              'label_id': '*',
@@ -590,6 +590,10 @@ class MeetingConfigDescriptor(Descriptor):
             {'name': "providedFollowUp",
              'view': "python: item.may_view_follow_up()",
              'edit': "python: item.may_edit_follow_up('providedFollowUp')"},
+            {'name': "groupsInChargeNotes",
+             'view': "python: tool.isManager(cfg) or tool.user_is_in_org(org_uid=[item.getProposingGroup()] + "
+                "item.getGroupsInCharge())",
+             'edit': "python: tool.isManager(cfg) or tool.user_is_in_org(org_uid=item.getGroupsInCharge())"},
         )
         self.votesResultTALExpr = ''